添加script setup loader配置
