+++ packages/webpack-plugin/lib/selector.js
@@ -1,6 +1,5 @@
 const parseComponent = require('./parser')
 const parseRequest = require('./utils/parse-request')
-const { compileScriptSetup } = require('./compiler-script-setup')
 
 module.exports = function (content) {
   this.cacheable()
@@ -25,10 +24,6 @@ module.exports = function (content) {
   let part = parts[type]
   if (Array.isArray(part)) {
     part = part[index]
-  }
-  if (part && part.attrs && part.attrs.setup) {
-    // script setup
-    part.content = compileScriptSetup(part, ctorType, filePath).content
   }
   if (!part) {
     let content = ''
